Clean up sum Ops
We now have two stats Ops:
1. Computes sum for RAIs using LoopBuilder
2. Computes sum for Iterables using enhanced for loop

Since the LoopBuilder op is faster (due to multi-threading), it is
higher priority

+++ imagej/imagej-ops2/src/test/java/net/imagej/ops2/stats/StatisticsTest.java
@@ -171,7 +171,7 @@ public class StatisticsTest extends AbstractOpTest {
 
 	@Test
 	public void testSum() {
-		final DoubleType sum = new DoubleType();
+		final FloatType sum = new FloatType();
 		ops.op("stats.sum").input(randomlyFilledImg).output(sum).compute();
 		Assertions.assertEquals(1277534.0, sum.getRealDouble(), 0.00001d, "Sum");
 	}
